move laws tests to .jvm dir.
