+++ fundamentals/fundamentals-bio/.jvm/src/test/scala/izumi/functional/bio/test/MiniBIOLawsTest.scala
@@ -1,11 +1,11 @@
-package izumi.fundamentals.bio
+package izumi.functional.bio.test
 
 import cats.effect.Sync
 import cats.effect.laws.SyncLaws
 import cats.effect.laws.discipline.SyncTests
 import izumi.functional.bio.catz._
+import izumi.functional.bio.env.MiniBIOEnv
 import izumi.functional.bio.impl.MiniBIO
-import izumi.fundamentals.bio.env.MiniBIOEnv
 
 class MiniBIOLawsTest extends CatsLawsTestBase with MiniBIOEnv {
     val syncTests = new SyncTests[MiniBIO[Throwable, ?]] {
